feat: practitioner appointments
